Fixes #39 - Unable to follow symlinks.
Turns out that we were naively going up a directoryon a bash script.

However, we should first find out where the bash script *really* is located,
then go up a directory and take the real directory of that to find the installation.

This is friendlier to our debian/rpm start scripts.  Thanks to @muuki88 for diagnosing
and proposing an alternative solution.