Support Docker volumes.
Docker has a feature where it can import or export "volumes", or subtrees of
the filesystem which can be shared in different ways between Docker containers,
or between a container and the host machine.

This adds the command to the Dockerfile which sets up the knowledge inside
Docker. Also, if a directory is created which does not exist, it will be owned
by root. This would be unfortunate if, for example, the exported directory
was the logs/ directory for Play. So the directories are created beforehand.