Added launcher jar plugin
The launcher jar plugin generates a jar with a classpath and main class
in the manifest.

Also refactored the classpath jar plugin to use standard sbt ways of
building jars, and to depend on the scriptClasspath so that projects
that modify it (such as Play) will still work with it.