An Vagrant+Ansible automated test project
This provides a new test project in the `test-project-play-rpm` directory. It is the default Play 2.3.x sample project packaged as an RPM.

The build and test environment uses a Centos-6.6 VM, managed with Vagrant and provisioned with Ansible (see the vagrant file and `provisioning/site.yml`).

The the package building and tests are also orchestrated with Ansible (see `provisioning/test.yml`).

The tests run performs the following:
- compile and build rpm package
- install the built rpm
- check that server is running and responding after install
- check that service can restart
- and is responding after restart
- uninstall the rpm