Adding 'dockerBuild{Command, Options}' and renaming 'dockerTag'
* dockerBuildCommand
The 'dockerBuildCommand' let users customize the command for building
Docker images.
The current default is 'docker build [dockerBuildOptions] .'.

* dockerBuildOptions
Furthermore 'dockerBuildOptions' defines which build options are used.
The current default is '--force-rm -t [dockerAlias]' (which is expanded
if 'dockerUpdateLatest' is set to 'true'.

This results in an execution of 'docker build --force-rm -t
[dockerAlias] .' (if 'dockerUpdateLatest' is 'false').

* dockerTag being renamed to dockerAlias
The 'dockerTag' setting is renamed to 'dockerAlias' as this better
reflects the setting. Docker aliases consist of a registry path,
username, name and tag ([REGISTRY_HOST/][USERNAME/]NAME[:TAG]).
To not confuse the Docker image tag with this setting (as it takes the
full alias name), the setting is renamed.