DockerAlias should have version scoped to Docker (#905)
When the SBT project version follows semantic versioning, and includes build metadata (e.g. 1.0.0+abcdef1) the Docker tag and publish is broken. The most reasonable way to handle this is to modify the build.sbt to adjust the version in Docker scope.

Right now, the DockerAlias pulls the version from the project scope.

-- Modified build.sbt --
(version in Docker) := version.value.split('+')(0)
-- In SBT console --
[sample-server] $ show dockerAlias
[info] DockerAlias(Some(some.registry),None,aergo-server,Some(1.0.0+abcdef1))
[sample-server] $ show docker:version
[info] 1.0.0
-- In SBT console after fix --
[sample-server] $ show dockerAlias
[info] DockerAlias(Some(some.registry),None,aergo-server,Some(1.0.0))
[sample-server] $ show docker:version
[info] 1.0.0