Only set executable permission on staged files if required and not set
When staging files, only set the target executable permission if the source file is executable and the target file is not executable.

It can be problematic if setting the executable permission on a target file multiple times if using inotify to watch staged files as inotify events will be triggered even when the permission is already set. This fix ensures the executable permission is only set once, to avoid unnecessary inotify events.