Fix multi-platform builds with multiple aliases.
Multi-platform builds append a `-t` argument for each alias to the
`dockerBuildOptions`, but the `publishTask` function was ALWAYS
iterating over the aliases even if it was a multi-platform build.

The consequence was attempts to buildx and push the tags multiple times.
In some environments this leads to build failures because you can't push
the same tag twice.

This change simply does not iterate over the aliases for multiplatform builds.